backend/pyproject.toml
[project]
name = "quantraider-backend"
version = "0.1.0"
description = "Quantraider backend (FastAPI + MongoDB)"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "motor>=3.6.0",
  "pymongo>=4.6.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib>=1.7.4",
  "argon2-cffi>=23.1.0",
  "python-multipart>=0.0.9",
]

[tool.uvicorn]
factory = false

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
